YKI(Frontend): Refactored Cypress tests
