+++ frontend/packages/yki/public/package.json
@@ -23,7 +23,7 @@
     "yki:start:dev-server": "yarn g:webpack serve --config webpack.config.js --env proxy=http://localhost:8083",
     "yki:start:docker-compose": "yarn g:webpack serve --config webpack.config.js --env proxy=http://yki-backend:8080 --env docker",
     "yki:stylelint": "yarn g:stylelint --fix \"./src/**/*.scss\"",
-    "yki:test:cypress": "TZ=Europe/Helsinki yarn g:cypress run --spec=\"./src/tests/cypress/integration/public/*.ts\"",
+    "yki:test:cypress": "TZ=Europe/Helsinki yarn g:cypress run",
     "yki:test:cypress:open": "TZ=Europe/Helsinki yarn g:cypress open",
     "yki:test:jest": "yarn g:jest",
     "yki:tslint": "yarn g:tsc --pretty --noEmit"
