+++ frontend/packages/yki/public/src/tests/cypress/integration/public_registration_page.spec.ts
