+++ frontend/packages/yki/public/src/tests/cypress/integration/public_registration_page_mobile.spec.ts
